requirements.txt
# FurniQuote AI · Python 依赖
# 阶段 1（骨架）只装核心，外部 SDK 在各模块开发时按需追加

# Web 框架
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# 配置与日志
PyYAML>=6.0.1
python-dotenv>=1.0.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# 远期模块（注释保留，待开发时启用）
# sqlalchemy>=2.0.0
# alembic>=1.13.0
# redis>=5.0.0
# chromadb>=0.4.22
# dashscope>=1.14.0
# anthropic>=0.20.0
# openpyxl>=3.1.0
# reportlab>=4.0.0
# Pillow>=10.0.0
